Development

+++ src/application/domain/barang/barang.domain.js
@@ -1,9 +1,10 @@
 const Barang = require('../../../model/barang');
+
 const {
-    searchKategori
+    getKategori
 } = require('../kategori/kategori.domain');
 const {
-    searchLokasi
+    getLokasi
 } = require('../lokasi/lokasi.domain');
 
 async function searchBarang(namaBarang) {
@@ -27,24 +28,70 @@ async function searchBarang(namaBarang) {
 }
 
 async function fetchBarang() {
-    const response = await Barang.find({});
+    const response = await Barang.find({}).populate(["kategori", "lokasi", "penggunaSaatIni"]);
+    const dataBaru = response.map(function (d) {
+        let pengguna;
+        if (d.penggunaSaatIni !== null) {
+            pengguna = d.penggunaSaatIni.namaKaryawan
+        } else {
+            pengguna = null;
+        }
+        return {
+            _id: d._id,
+            kodeBarang: d.kodeBarang,
+            namaBarang: d.namaBarang,
+            kategori: d.kategori.namaKategori,
+            tanggalPerolehan: d.tanggalPerolehan,
+            hargaPerolehan: d.hargaPerolehan,
+            masaGuna: d.masaGuna,
+            kondisi: d.kondisi,
+            lokasi: d.lokasi.namaLokasi,
+            penggunaSaatIni: pengguna,
+            createdAt: d.createdAt,
+            updatedAt: d.updatedAt,
+        }
+    });
     return {
         status: 200,
-        data: response,
+        data: dataBaru,
     };
 }
 
 async function getBarang(id) {
     try {
-        const response = await Barang.findOne({
+        const response = await Barang.find({
             _id: id
-        });
+        }).populate(["kategori", "lokasi", "penggunaSaatIni"]);
+
         if (response === null) {
             throw new Error();
         }
+        const dataBaru = response.map(function (d) {
+            let pengguna;
+            if (d.penggunaSaatIni !== null) {
+                pengguna = d.penggunaSaatIni.namaKaryawan
+            } else {
+                pengguna = null;
+            }
+            return {
+                _id: d._id,
+                kodeBarang: d.kodeBarang,
+                namaBarang: d.namaBarang,
+                kategori: d.kategori.namaKategori,
+                tanggalPerolehan: d.tanggalPerolehan,
+                hargaPerolehan: d.hargaPerolehan,
+                masaGuna: d.masaGuna,
+                kondisi: d.kondisi,
+                lokasi: d.lokasi.namaLokasi,
+                penggunaSaatIni: pengguna,
+                createdAt: d.createdAt,
+                updatedAt: d.updatedAt,
+            }
+        });
+
         return {
             status: 200,
-            data: response
+            data: dataBaru
         };
     } catch (error) {
         return {
@@ -60,10 +107,9 @@ async function createBarang(dataBarang) {
     currentTime.setMinutes(currentTime.getMinutes() + offset);
     dataBarang.createdAt = currentTime;
     dataBarang.updatedAt = currentTime;
-    dataBarang.penggunaSaatIni = "-";
     dataBarang.kodeBarang = generateRandomString(7);
-    const kategori = await searchKategori(dataBarang.kategori);
-    const lokasi = await searchLokasi(dataBarang.lokasi);
+    const kategori = await getKategori(dataBarang.kategori);
+    const lokasi = await getLokasi(dataBarang.lokasi);
     try {
         if (kategori.status !== 200) {
             throw new Error('Kategori tersebut tidak ada');
@@ -71,9 +117,9 @@ async function createBarang(dataBarang) {
         if (lokasi.status !== 200) {
             throw new Error('Lokasi penyimpanan tersebut tidak ada');
         }
-        const create = new Barang(dataBarang);
-        const data = await create.save();
-        if (!data) {
+        dataBarang.penggunaSaatIni = null;
+        await Barang.create(dataBarang);
+        if (!Barang) {
             throw new Error("Gagal menambahkan Barang")
         }
         return {
@@ -92,19 +138,18 @@ async function updateBarang(id, dataBarang) {
     const offset = 420;
     currentTime.setMinutes(currentTime.getMinutes() + offset);
     dataBarang.updatedAt = currentTime;
-    dataBarang.penggunaSaatIni = "-";
-    const kategori = await searchKategori(dataBarang.kategori);
-    const lokasi = await searchLokasi(dataBarang.lokasi);
     try {
-        if (kategori.status !== 200) {
-            throw new Error('Kategori tersebut tidak ada');
+        if (dataBarang.hasOwnProperty('kategori') && dataBarang.hasOwnProperty('lokasi')) {
+            const kategori = await getKategori(dataBarang.kategori);
+            const lokasi = await getLokasi(dataBarang.lokasi);
+            if (kategori.status !== 200) {
+                throw new Error('Kategori tersebut tidak ada');
+            }
+            if (lokasi.status !== 200) {
+                throw new Error('Lokasi penyimpanan tersebut tidak ada');
+            }
         }
-        if (lokasi.status !== 200) {
-            throw new Error('Lokasi penyimpanan tersebut tidak ada');
-        }
-        const update = await Barang.updateOne({
-            _id: id
-        }, dataBarang);
+        const update = await Barang.findByIdAndUpdate(id, dataBarang);
         if (!update) {
             throw new Error('Gagal memperbarui Barang', );
         }
@@ -121,17 +166,22 @@ async function updateBarang(id, dataBarang) {
 }
 async function deleteBarang(id) {
     try {
-        await Barang.deleteOne({
-            _id: id
-        });
-        return {
-            status: 200,
-            message: 'Berhasil menghapus Barang'
-        };
+        const pengguna = await getBarang(id);
+        if(pengguna.data[0].penggunaSaatIni === null){
+            await Barang.deleteOne({
+                _id: id
+            });
+            return {
+                status: 200,
+                message: 'Berhasil menghapus Barang'
+            };
+        }else{
+            throw new Error(`Tidak dapat menghapus. Barang ini sedang digunakan oleh ${pengguna.data[0].penggunaSaatIni}`);
+        }
     } catch (error) {
         return {
             status: 500,
-            message: 'Gagal menghapus Barang'
+            message: error.message
         };
     }
 }
